Allow custom kubeconfig location for kubectl
Every time the kube-solo VM is brought up, the kubectl wrapper is
regenerated.  This means that any customizations you make to this file,
like the location of the kubectl configuration, is overwritten.  This
commit will allow people to use an environment variable to point where
the kubectl wrapper finds the kubectl configuration file instead of us
having to update the wrapper each time the VM is brought up.